backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-doctr[torch]==0.10.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
spacy>=3.7.0
Pillow>=10.0.0
pdf2image>=1.17.0
pydantic>=2.0.0
numpy>=1.26.0
